+++ src/adc/writer/speech/__init__.py
@@ -2,4 +2,5 @@ from ._adams import AdamsSpeechWriter
 from ._commonvoice import CommonVoiceSpeechWriter
 from ._festvox import FestVoxSpeechWriter
 from ._hf_audiofolder import HuggingFaceAudioFolderSpeechWriter
+from ._piper import PiperSpeechWriter
 from ._txt import TxtSpeechWriter