+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - `to-commonvoice-sp` writer now creates output directories automatically
 - `from-commonvoice-sp` reader now adds the additional columns from its .tsv file into meta-data
+- added `to-piper-sp` writer (https://github.com/rhasspy/piper/blob/master/TRAINING.md#dataset-format)
 
 
 0.1.0 (2025-10-31)